Make it to be compatible with PyPI
