+++ compile.sh
@@ -3,7 +3,11 @@
 echo "Compile OctaDist using PyInstaller"
 echo "=================================="
 
-pyinstaller main.py --onefile --version-file version.txt -i molecule.ico --name OctaDist-2.2-Win-x86-64 --distpath "../OctaDist-Executable"
+pyinstaller -F -w -i molecule.ico --version-file version.txt -n OctaDist-2.5-Win-x86-64
 
 # Optional
-# --windowed (turn of window console)
+# --onefile		or -F	(compile as a standalone executable)
+# --windowed 	or -w	(turn off window console)
+# -i 					(set program icon)
+# --version-file		(set program version file)
+# --name 		or -n 	(set program name)
