+++ example-input/[Fe(1-bpp)2]I0.5[I3]1.xyz
@@ -1,0 +1,12 @@
+7
+
+Fe                11.01876829   13.04878092    9.30523187
+N                 12.08686829   14.74288092    8.55033187
+N                 10.40276829   14.76468092   10.50093187
+N                 12.66486829   12.29858092    8.08173187
+N                  9.60636829   11.41678092    9.23563187
+N                 11.56446829   11.68348092   10.90363187
+N                  9.47066829   13.52208092    7.76973187
+
+
+
