+++ example-input/[Fe(1-bpp)2][BF4]2-HS-Full.xyz
@@ -1,0 +1,54 @@
+51
+[Fe(1-bpp)2][BF4]2-HS-Full
+Fe                 4.90490000    6.91350000   14.24800000
+N                  4.98220000    6.87650000   12.11090000
+N                  2.95710000    7.67330000   13.54390000
+N                  4.67140000    6.74120000   16.36850000
+N                  6.85350000    6.08640000   13.70170000
+N                  5.68300000    8.77920000   15.10820000
+N                  4.10760000    4.89840000   14.64310000
+C                  3.95110000    7.34120000   11.39040000
+C                  3.96790000    7.39090000    9.99990000
+H                  3.24140000    7.70930000    9.51520000
+C                  5.12240000    6.94150000    9.36990000
+H                  5.17300000    6.96380000    8.44210000
+C                  6.19930000    6.45960000   10.09670000
+H                  6.97470000    6.16260000    9.67910000
+C                  6.06580000    6.44250000   11.47410000
+N                  2.85780000    7.75980000   12.17480000
+C                  1.78900000    8.12950000   14.00140000
+H                  1.56160000    8.18350000   14.90180000
+C                  0.94950000    8.51730000   12.94770000
+H                  0.09230000    8.87160000   13.01670000
+C                  1.64550000    8.26900000   11.80190000
+H                  1.35200000    8.41800000   10.93140000
+N                  7.06320000    5.96320000   12.35070000
+C                  7.90470000    5.49240000   14.26540000
+H                  8.04400000    5.42910000   15.18200000
+C                  8.77020000    4.97630000   13.29880000
+H                  9.56660000    4.51930000   13.44740000
+C                  8.20860000    5.28020000   12.08800000
+H                  8.54430000    5.06110000   11.24970000
+C                  5.03250000    7.74860000   17.16910000
+C                  4.88910000    7.70330000   18.55180000
+H                  5.14250000    8.41630000   19.09330000
+C                  4.34870000    6.54350000   19.08240000
+H                  4.23050000    6.47760000   20.00250000
+C                  3.98030000    5.48050000   18.27800000
+H                  3.62870000    4.69640000   18.63580000
+C                  4.16090000    5.64140000   16.91650000
+N                  5.59200000    8.84510000   16.47680000
+C                  6.30310000    9.90980000   14.75750000
+H                  6.50150000   10.15040000   13.88200000
+C                  6.61580000   10.68700000   15.88340000
+H                  7.05020000   11.50870000   15.89870000
+C                  6.14730000    9.98350000   16.94930000
+H                  6.19860000   10.24020000   17.84100000
+N                  3.82500000    4.65280000   15.95530000
+C                  3.73140000    3.80030000   13.98740000
+H                  3.81770000    3.68050000   13.07000000
+C                  3.18990000    2.85710000   14.85780000
+H                  2.84890000    2.02000000   14.64070000
+C                  3.26630000    3.43140000   16.11790000
+H                  2.98720000    3.05050000   16.91840000
+
