+++ example-input/[Fe(1-bpp)2][BF4]2-LS-Full.xyz
@@ -1,0 +1,54 @@
+51
+[Fe(1-bpp)2][BF4]2-LS-Full
+Fe                 4.06740000    7.20400000   13.61170000
+N                  4.30330000    7.37500000   11.72920000
+N                  2.31840000    8.01650000   13.11520000
+N                  5.88220000    6.44610000   13.43120000
+N                  3.83260000    6.97150000   15.49260000
+N                  4.80550000    8.93180000   14.27160000
+N                  3.30020000    5.38280000   13.63160000
+C                  3.32570000    7.88480000   10.96510000
+C                  3.43510000    8.00970000    9.58920000
+H                  2.73580000    8.37830000    9.06250000
+C                  4.63370000    7.56230000    9.01840000
+H                  4.74970000    7.63420000    8.07790000
+C                  5.65940000    7.01570000    9.78100000
+H                  6.46920000    6.71470000    9.38710000
+C                  5.43880000    6.93450000   11.15020000
+N                  2.22100000    8.25180000   11.75890000
+C                  1.15110000    8.42450000   13.62940000
+H                  0.92390000    8.38350000   14.55110000
+C                  0.30430000    8.92410000   12.61430000
+H                 -0.57080000    9.27650000   12.72400000
+C                  1.00220000    8.79660000   11.43760000
+H                  0.69950000    9.03960000   10.57050000
+N                  6.30750000    6.38700000   12.11440000
+C                  6.84280000    5.85930000   14.13390000
+H                  6.83070000    5.75320000   15.07810000
+C                  7.88870000    5.41110000   13.28930000
+H                  8.68530000    4.96540000   13.55190000
+C                  7.51650000    5.75150000   12.01650000
+H                  8.00390000    5.58040000   11.21970000
+C                  4.21100000    7.93100000   16.34690000
+C                  4.08120000    7.80100000   17.72300000
+H                  4.34890000    8.49130000   18.31840000
+C                  3.53430000    6.59920000   18.18980000
+H                  3.42420000    6.47340000   19.12430000
+C                  3.14630000    5.58390000   17.31980000
+H                  2.78150000    4.76780000   17.64000000
+C                  3.31780000    5.81990000   15.96650000
+N                  4.75750000    9.02590000   15.65690000
+C                  5.42180000   10.03780000   13.86440000
+H                  5.59430000   10.25590000   12.95560000
+C                  5.78680000   10.84870000   14.97180000
+H                  6.23910000   11.68350000   14.94680000
+C                  5.35100000   10.17460000   16.08810000
+H                  5.44530000   10.45600000   16.99090000
+N                  3.00400000    4.94490000   14.90920000
+C                  3.00790000    4.35980000   12.83430000
+H                  3.11840000    4.36580000   11.89070000
+C                  2.51300000    3.26320000   13.57170000
+H                  2.23080000    2.42240000   13.23090000
+C                  2.52450000    3.66700000   14.88790000
+H                  2.25040000    3.15370000   15.63790000
+
