+++ example-input/[Fe(1-bpp)2][ClO4]2.xyz
@@ -1,0 +1,11 @@
+7
+
+Fe                 3.41725012    7.16918286    6.68113278
+N                  3.27345012    7.62638286    8.79513278
+N                  2.28705012    5.56418286    7.63143278
+N                  4.88045012    8.76048286    7.13443278
+N                  3.56115012    7.62638286    4.56713278
+N                  4.54745012    5.56418286    5.73083278
+N                  1.95405012    8.76048286    6.22773278
+
+
