+++ example-input/[Fe(1-bpp)2][I3]2.xyz
@@ -1,0 +1,10 @@
+7
+Bis[2,6-di(pyrazol-1-yl)pyridine]iron(II) triiodide  [Fe(1-bpp)2][I3]2
+Fe                14.68572000   18.49228000    6.66716000
+N                 16.26114000   18.54903000    8.20527000
+N                 14.44181000   20.59400000    6.21555000
+N                 16.09157000   18.96170000    5.02956000
+N                 13.04897000   19.25464000    7.93122000
+N                 14.86476000   16.48821000    7.43379000
+N                 13.37473000   17.23453000    5.45099000
+
