+++ example-input/[Fe(1-bpp)2][PF6]2.xyz
@@ -1,0 +1,12 @@
+7
+
+Fe                -0.77115220   -0.45477076    4.26988462
+N                 -0.97515220   -0.94007076    6.37758462
+N                 -1.91935220    1.15582924    5.23018462
+N                  0.65104780   -2.08877076    4.74698462
+N                 -0.56705220   -0.94007076    2.16228462
+N                  0.37704780    1.15582924    3.30968462
+N                 -2.19325220   -2.08877076    3.79288462
+
+
+
