+++ example-input/[Fe(1-bpp)2][SbF6]2.xyz
@@ -1,0 +1,11 @@
+7
+
+Fe                 1.89703179   10.07980056    3.20191719
+N                  1.65873179    9.59830056    5.30801719
+N                  0.74803179   11.69410056    4.14441719
+N                  3.31073179    8.43950056    3.70331719
+N                  2.13543179    9.59830056    1.09581719
+N                  3.04613179   11.69410056    2.25941719
+N                  0.48333179    8.43950056    2.70051719
+
+
