--- test/Fe-distorted-octa.xyz
@@ -1,9 +1,0 @@
-7
-From Excel file
-Fe  6.251705    9.063211    5.914842
-N   8.15961     9.066456    5.463087
-N   6.749414    10.457551   7.179682
-N   5.709997    10.492955   4.658257
-N   4.350474    9.106286    6.356091
-O   5.789096    7.796326    4.611355
-O   6.686381    7.763872    7.209699
