--- test/Fe-very-distorted-octa.xyz
@@ -1,9 +1,0 @@
-7
-Cartesian coordinate
- Fe        18.268051000    11.289120000     2.565804000
-  O        19.823874000    10.436314000     1.381569000
-  O        19.074466000     9.706294000     3.743576000
-  N        17.364238000    10.733354000     0.657318000
-  N        16.149538000    11.306661000     2.913619000
-  N        18.599941000    12.116308000     4.528988000
-  N        18.364987000    13.407634000     2.249608000
