--- test/Perfect-octahedron.xyz
@@ -1,9 +1,0 @@
-7
-
-Fe                 0.20069808    0.70680627    0.00000000
-O                  1.66069808    0.70680627    0.00000000
-O                  0.20069808    2.16680627    0.00000000
-O                  0.20069808    0.70680627    1.46000000
-O                 -1.25930192    0.70680627    0.00000000
-O                  0.20069808   -0.75319373    0.00000000
-O                  0.20069808    0.70680627   -1.46000000
