+++ test_octadist.py
@@ -1,0 +1,6 @@
+import octadist
+
+A = [[]]
+
+octadist.calc.calc_all(self, A)
+
